requirements.txt
chess==1.9.1
PyYAML==6.0
numpy==1.22.4